added summary-report apis

+++ backend/app/controllers/categoryController.js
@@ -1,4 +1,5 @@
 const Category = require('../models/categoryModel');
+const Expense = require('../models/expenseModel')
 const categoryController = {};
 
 categoryController.addCategory = async (req, res) => {
@@ -38,6 +39,15 @@ categoryController.updateCategory = async (req, res) => {
 categoryController.removeCategory = async (req, res) => {
     try {
         const categoryId = req.params.id;
+        const associatedExpenses = await Expense.find({ category: categoryId }).populate('category', 'name');
+
+        if (associatedExpenses.length > 0) {
+            const category = await Category.findById(categoryId);
+            if (!category) {
+                return res.status(404).json({ error: 'Category not found' });
+            }
+            return res.status(400).json({ error: `Cannot delete the category named "${category.name}" as it is associated with one or more expenses.` });
+        }
         const category = await Category.findByIdAndDelete(categoryId);
         if (!category) {
             return res.status(404).json({ error: 'Category not found' });
