+++ backend/app/controllers/userController.js
@@ -1,4 +1,6 @@
 const User = require('../models/userModel')
+const Expense = require('../models/expenseModel')
+
 const userController = {}
 userController.addUser =  async (req, res) => {
     try{
@@ -39,6 +41,16 @@ userController.updateUser = async (req, res) => {
 userController.removeUser = async (req, res) => {
     try {
         const userId = req.params.id;
+        const associatedExpenses = await Expense.find({ user: userId });
+
+        if (associatedExpenses.length > 0) {
+            const user = await User.findById(userId);
+            if (!user) {
+                return res.status(404).json({ error: 'User not found' });
+            }
+            return res.status(400).json({ error: `Cannot delete the user named "${user.name}"as the user is associated with one or more expenses.` });
+        }
+
         const user = await User.findByIdAndDelete(userId);
         if (!user) {
             return res.status(404).json({ error: 'User not found' });
